Extract embedded type to new file
